combine the Doclinks and Codelinks modules, this separation is not worthwhile
