+++ Sources/LinkResolution/Codelinks/CodelinkResolutionError.Note.swift
@@ -1,4 +1,4 @@
-import Codelinks
+import UCF
 import SourceDiagnostics
 
 extension CodelinkResolutionError
