+++ Sources/LinkResolution/Codelinks/CodelinkResolutionError.swift
@@ -1,4 +1,4 @@
-import Codelinks
+import UCF
 import SourceDiagnostics
 
 @frozen public
