+++ Sources/LinkResolution/Codelinks/CodelinkResolver.Scope.swift
