+++ Sources/LinkResolution/Codelinks/CodelinkResolver.Table.swift
@@ -1,4 +1,4 @@
-import Codelinks
+import UCF
 import FNV1
 import LexicalPaths
 import Symbols
