+++ Sources/LinkResolution/Codelinks/CodelinkResolver.swift
@@ -1,4 +1,4 @@
-import Codelinks
+import UCF
 import Symbols
 import Unidoc
