+++ Sources/LinkResolution/Doclinks/DoclinkResolver.Scope.swift
