+++ Sources/LinkResolution/Doclinks/DoclinkResolver.swift
@@ -1,4 +1,4 @@
-import Doclinks
+import UCF
 import Symbols
 
 @frozen public
