+++ Sources/MarkdownSemantics/Sema/Markdown.SemanticSections.swift
@@ -30,6 +30,12 @@ extension Markdown
 }
 extension Markdown.SemanticSections
 {
+    // public
+    // func anchors() -> [UCF.AnchorIdentifier: String]
+    // {
+
+    // }
+
     @inlinable public
     var isEmpty:Bool
     {
