+++ Sources/SymbolGraphLinker/Resolution/SSGC.OutlineResolutionScopes.swift
@@ -1,5 +1,4 @@
-import CodelinkResolution
-import DoclinkResolution
+import LinkResolution
 import Symbols
 
 extension SSGC
