+++ Sources/SymbolGraphLinker/Resolution/SSGC.Outliner.Warning.swift
@@ -1,5 +1,5 @@
-import Doclinks
 import SourceDiagnostics
+import UCF
 
 extension SSGC.Outliner
 {
