+++ Sources/UCF/Codelinks/Codelink.Path.IdentifierHead.swift
