+++ Sources/UCF/Codelinks/Codelink.Path.OperatorHead.swift
