+++ Sources/UCF/Codelinks/Codelink.Suffix.swift
