+++ Sources/UCFTests/Doclink (ext).swift
@@ -1,5 +1,5 @@
-import Doclinks
 import Testing_
+import UCF
 
 extension Doclink
 {
