+++ Sources/UCFTests/Main.swift
@@ -6,6 +6,7 @@ enum Main:TestMain
     static
     let all:[any TestBattery.Type] =
     [
-        Parsing.self,
+        ParseCodelink.self,
+        ParseDoclink.self,
     ]
 }
