+++ Sources/UnidocLinker/Resolver/Codelink (ext).swift
@@ -1,6 +1,5 @@
-import Codelinks
-import Doclinks
 import SymbolGraphs
+import UCF
 
 extension Codelink
 {
